Keep standard descriptor regressions in existing protobuf tests
